rag-chain/requirements.txt
sentence-transformers
numpy
langgraph==1.0.2
google-generativeai
